mito-ai: refactor
